Move getUserInfo from App --> Home
